Move AttributeInput
